+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- New `RemoveGeneratedThumbnails` script
+
 ### Updated
 - Update dspace-api dependency to 7.6.3
 - Update gson dependency to 2.11.0 to match dspace-api